+++ src/dialogs/more-info/components/lights/dialog-light-color-favorite.ts
@@ -1,16 +1,25 @@
 import { mdiClose } from "@mdi/js";
 import { css, CSSResultGroup, html, LitElement, nothing } from "lit";
 import { customElement, property, state } from "lit/decorators";
+import { classMap } from "lit/directives/class-map";
 import { fireEvent } from "../../../../common/dom/fire_event";
 import "../../../../components/ha-button";
 import "../../../../components/ha-dialog";
 import "../../../../components/ha-dialog-header";
-import { EntityRegistryEntry } from "../../../../data/entity_registry";
-import { LightColor } from "../../../../data/light";
+import "../../../../components/ha-icon-button-toggle";
+import type { EntityRegistryEntry } from "../../../../data/entity_registry";
+import {
+  LightColor,
+  LightColorMode,
+  LightEntity,
+  lightSupportsColor,
+  lightSupportsColorMode,
+} from "../../../../data/light";
 import { haStyleDialog } from "../../../../resources/styles";
-import { HomeAssistant } from "../../../../types";
-import "./light-color-picker";
-import { LightColorFavoriteDialogParams } from "./show-dialog-light-color-favorite";
+import type { HomeAssistant } from "../../../../types";
+import type { LightColorFavoriteDialogParams } from "./show-dialog-light-color-favorite";
+
+type LightPickerMode = "color_temp" | "color";
 
 @customElement("dialog-light-color-favorite")
 class DialogLightColorFavorite extends LitElement {
@@ -22,11 +31,16 @@ class DialogLightColorFavorite extends LitElement {
 
   @state() _color?: LightColor;
 
+  @state() private _mode?: LightPickerMode;
+
+  @state() private _modes: LightPickerMode[] = [];
+
   public async showDialog(
     dialogParams: LightColorFavoriteDialogParams
   ): Promise<void> {
     this._entry = dialogParams.entry;
     this._dialogParams = dialogParams;
+    this._updateModes(dialogParams.defaultMode);
     await this.updateComplete;
   }
 
@@ -37,8 +51,41 @@ class DialogLightColorFavorite extends LitElement {
     fireEvent(this, "dialog-closed", { dialog: this.localName });
   }
 
+  private _updateModes(defaultMode?: LightPickerMode) {
+    const supportsTemp = lightSupportsColorMode(
+      this.stateObj!,
+      LightColorMode.COLOR_TEMP
+    );
+
+    const supportsColor = lightSupportsColor(this.stateObj!);
+
+    const modes: LightPickerMode[] = [];
+    if (supportsColor) {
+      modes.push("color");
+    }
+    if (supportsTemp) {
+      modes.push("color_temp");
+    }
+
+    this._modes = modes;
+    this._mode =
+      defaultMode ??
+      (this.stateObj!.attributes.color_mode
+        ? this.stateObj!.attributes.color_mode === LightColorMode.COLOR_TEMP
+          ? LightColorMode.COLOR_TEMP
+          : "color"
+        : this._modes[0]);
+  }
+
   private _colorChanged(ev: CustomEvent) {
     this._color = ev.detail;
+  }
+
+  get stateObj() {
+    return (
+      this._entry &&
+      (this.hass.states[this._entry.entity_id] as LightEntity | undefined)
+    );
   }
 
   private async _cancel() {
@@ -55,8 +102,16 @@ class DialogLightColorFavorite extends LitElement {
     this.closeDialog();
   }
 
+  private _modeChanged(ev): void {
+    const newMode = ev.currentTarget.mode;
+    if (newMode === this._mode) {
+      return;
+    }
+    this._mode = newMode;
+  }
+
   protected render() {
-    if (!this._entry) {
+    if (!this._entry || !this.stateObj) {
       return nothing;
     }
 
@@ -76,13 +131,55 @@ class DialogLightColorFavorite extends LitElement {
           ></ha-icon-button>
           <span slot="title">${this._dialogParams?.title}</span>
         </ha-dialog-header>
-        <light-color-picker
-          .hass=${this.hass}
-          entityId=${this._entry.entity_id}
-          .defaultMode=${this._dialogParams?.defaultMode}
-          @color-changed=${this._colorChanged}
-        >
-        </light-color-picker>
+        ${this._modes.length > 1
+          ? html`
+              <div class="modes">
+                ${this._modes.map(
+                  (value) =>
+                    html`
+                      <ha-icon-button-toggle
+                        border-only
+                        ?selected=${value === this._mode}
+                        .title=${this.hass.localize(
+                          `ui.dialogs.more_info_control.light.color_picker.mode.${value}`
+                        )}
+                        .ariaLabel=${this.hass.localize(
+                          `ui.dialogs.more_info_control.light.color_picker.mode.${value}`
+                        )}
+                        .mode=${value}
+                        @click=${this._modeChanged}
+                      >
+                        <span
+                          class="wheel ${classMap({ [value]: true })}"
+                        ></span>
+                      </ha-icon-button-toggle>
+                    `
+                )}
+              </div>
+            `
+          : nothing}
+        <div class="content">
+          ${this._mode === "color_temp"
+            ? html`
+                <light-color-temp-picker
+                  .hass=${this.hass}
+                  .stateObj=${this.stateObj}
+                  @color-changed=${this._colorChanged}
+                >
+                </light-color-temp-picker>
+              `
+            : nothing}
+          ${this._mode === "color"
+            ? html`
+                <light-color-rgb-picker
+                  .hass=${this.hass}
+                  .stateObj=${this.stateObj}
+                  @color-changed=${this._colorChanged}
+                >
+                </light-color-rgb-picker>
+              `
+            : nothing}
+        </div>
         <ha-button slot="secondaryAction" dialogAction="cancel">
           ${this.hass.localize("ui.common.cancel")}
         </ha-button>
@@ -101,12 +198,6 @@ class DialogLightColorFavorite extends LitElement {
           --dialog-content-padding: 0;
         }
 
-        light-color-picker {
-          display: flex;
-          flex-direction: column;
-          flex: 1;
-        }
-
         @media all and (max-width: 450px), all and (max-height: 500px) {
           ha-dialog {
             --dialog-surface-margin-top: 100px;
@@ -117,6 +208,39 @@ class DialogLightColorFavorite extends LitElement {
               28px 28px 0 0
             );
           }
+        }
+
+        .content {
+          display: flex;
+          flex-direction: column;
+          align-items: center;
+          justify-content: center;
+          padding: 24px;
+          flex: 1;
+        }
+        .modes {
+          display: flex;
+          flex-direction: row;
+          justify-content: flex-end;
+          padding: 0 24px;
+        }
+        .wheel {
+          width: 30px;
+          height: 30px;
+          flex: none;
+          border-radius: 15px;
+        }
+        .wheel.color {
+          background-image: url("/static/images/color_wheel.png");
+          background-size: cover;
+        }
+        .wheel.color_temp {
+          background: linear-gradient(
+            0,
+            rgb(166, 209, 255) 0%,
+            white 50%,
+            rgb(255, 160, 0) 100%
+          );
         }
       `,
     ];
